HW-3 Main UI MVP 적용
