+++ jskim5923/ArchitectureStudy/app/src/main/java/com/jskim5923/architecturestudy/main/MainActivity.kt
@@ -4,19 +4,17 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.jskim5923.architecturestudy.R
 import com.jskim5923.architecturestudy.adapter.ViewPagerAdapter
-import com.jskim5923.architecturestudy.extension.getCoinCurrency
-import com.jskim5923.architecturestudy.model.data.source.Repository
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.disposables.CompositeDisposable
-import io.reactivex.rxkotlin.plusAssign
-import io.reactivex.schedulers.Schedulers
+import com.jskim5923.architecturestudy.main.MainContract
+import com.jskim5923.architecturestudy.main.MainPresenter
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity() {
-    private val compositeDisposable = CompositeDisposable()
-
+class MainActivity : AppCompatActivity(), MainContract.View {
     private val viewpagerAdapter by lazy {
         ViewPagerAdapter(supportFragmentManager)
+    }
+
+    private val presenter by lazy {
+        MainPresenter(this)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -28,27 +26,15 @@ class MainActivity : AppCompatActivity() {
             adapter = viewpagerAdapter
         }
 
-        loadData()
+        presenter.loadMarketList()
     }
 
-    private fun loadData() {
-        compositeDisposable += Repository.getMarketList()
-            .subscribeOn(Schedulers.io())
-            .map { marketList ->
-                marketList.map {
-                    it.market.getCoinCurrency()
-                }.distinct()
-            }
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribe({
-                viewpagerAdapter.updateItem(it)
-            }, {
-                it.printStackTrace()
-            })
+    override fun updateViewpager(marketList: List<String>) {
+        viewpagerAdapter.updateItem(marketList)
     }
 
     override fun onDestroy() {
-        compositeDisposable.clear()
+        presenter.disposableClear()
         super.onDestroy()
     }
 }
