+++ jskim5923/ArchitectureStudy/app/src/main/java/com/jskim5923/architecturestudy/main/MainPresenter.kt
@@ -1,0 +1,33 @@
+package com.jskim5923.architecturestudy.main
+
+import com.jskim5923.architecturestudy.extension.getCoinCurrency
+import com.jskim5923.architecturestudy.model.data.source.Repository
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.rxkotlin.plusAssign
+import io.reactivex.schedulers.Schedulers
+
+class MainPresenter(private val view: MainContract.View) : MainContract.Presenter {
+    private val compositeDisposable = CompositeDisposable()
+
+    override fun loadMarketList() {
+        compositeDisposable += Repository.getMarketList()
+            .subscribeOn(Schedulers.io())
+            .map { marketList ->
+                marketList.map {
+                    it.market.getCoinCurrency()
+                }.distinct()
+            }
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe({
+                view.updateViewpager(it)
+            }, {
+                it.printStackTrace()
+            })
+    }
+
+    override fun disposableClear() {
+        compositeDisposable.clear()
+    }
+
+}
